requirements.txt
requests
googlesearch-python
googlesearch-python
flask
